🚚 Rename "Theme" module to "Design system" (#169)
Since the module contains more components and logic based on how the
application look and feel should behave, I believe that is better to
name it "Design System" in order to better reflect its importance and
content.
