+++ app/src/main/java/com/escodro/alkaa/presentation/home/Home.kt
@@ -38,11 +38,11 @@ import androidx.compose.ui.unit.dp
 import com.escodro.alkaa.model.HomeSection
 import com.escodro.category.presentation.CategoryBottomSheet
 import com.escodro.category.presentation.CategoryListSection
+import com.escodro.designsystem.AlkaaTheme
 import com.escodro.preference.presentation.PreferenceSection
 import com.escodro.search.presentation.SearchSection
 import com.escodro.task.presentation.add.AddTaskBottomSheet
 import com.escodro.task.presentation.list.TaskListSection
-import com.escodro.theme.AlkaaTheme
 import kotlinx.coroutines.launch
 
 /**
