+++ features/category/build.gradle
@@ -18,7 +18,7 @@ dependencies {
     implementation project(":libraries:core")
     implementation project(":features:category-api")
     implementation project(":domain")
-    implementation project(":libraries:theme")
+    implementation project(':libraries:designsystem')
 
     testImplementation project(":libraries:test")
     androidTestImplementation project(":libraries:test")
