+++ features/preference/build.gradle
@@ -14,7 +14,7 @@ android {
 
 dependencies {
     implementation project(":libraries:core")
-    implementation project(":libraries:theme")
+    implementation project(':libraries:designsystem')
 
     testImplementation project(":libraries:test")
     androidTestImplementation project(":libraries:test")
