+++ features/preference/src/androidTest/java/com/escodro/preference/PreferenceSectionTest.kt
@@ -3,8 +3,8 @@ package com.escodro.preference
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.test.platform.app.InstrumentationRegistry
+import com.escodro.designsystem.AlkaaTheme
 import com.escodro.preference.presentation.PreferenceSection
-import com.escodro.theme.AlkaaTheme
 import org.junit.Rule
 import org.junit.Test
