+++ features/search/src/main/java/com/escodro/search/presentation/Search.kt
@@ -37,11 +37,11 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.escodro.designsystem.AlkaaTheme
+import com.escodro.designsystem.components.AlkaaLoadingContent
+import com.escodro.designsystem.components.DefaultIconTextContent
 import com.escodro.search.R
 import com.escodro.search.model.TaskSearchItem
-import com.escodro.theme.AlkaaTheme
-import com.escodro.theme.components.AlkaaLoadingContent
-import com.escodro.theme.components.DefaultIconTextContent
 import org.koin.androidx.compose.getViewModel
 
 /**
