+++ features/task/src/main/java/com/escodro/task/presentation/list/TaskList.kt
@@ -32,15 +32,15 @@ import androidx.compose.ui.unit.dp
 import com.escodro.categoryapi.model.Category
 import com.escodro.categoryapi.presentation.CategoryListViewModel
 import com.escodro.categoryapi.presentation.CategoryState
+import com.escodro.designsystem.AlkaaTheme
+import com.escodro.designsystem.components.AddFloatingButton
+import com.escodro.designsystem.components.AlkaaLoadingContent
+import com.escodro.designsystem.components.DefaultIconTextContent
 import com.escodro.task.R
 import com.escodro.task.model.Task
 import com.escodro.task.model.TaskWithCategory
 import com.escodro.task.presentation.category.CategorySelection
 import com.escodro.task.presentation.detail.main.CategoryId
-import com.escodro.theme.AlkaaTheme
-import com.escodro.theme.components.AddFloatingButton
-import com.escodro.theme.components.AlkaaLoadingContent
-import com.escodro.theme.components.DefaultIconTextContent
 import kotlinx.coroutines.launch
 import org.koin.androidx.compose.getViewModel
 import java.util.Calendar
