+++ features/tracker/src/main/java/com/escodro/tracker/presentation/TrackerActivity.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import com.escodro.theme.AlkaaTheme
+import com.escodro.designsystem.AlkaaTheme
 import com.escodro.tracker.di.injectDynamicFeature
 import com.google.android.play.core.splitcompat.SplitCompat
