+++ libraries/designsystem/proguard-rules.pro
