+++ libraries/designsystem/src/main/java/com/escodro/designsystem/Color.kt
@@ -1,4 +1,4 @@
-package com.escodro.theme
+package com.escodro.designsystem
 
 import androidx.compose.ui.graphics.Color
