+++ libraries/designsystem/src/main/java/com/escodro/designsystem/Theme.kt
@@ -1,4 +1,4 @@
-package com.escodro.theme
+package com.escodro.designsystem
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
