+++ settings.gradle
@@ -12,5 +12,5 @@ include ':data:local'
 include ':libraries:core'
 include ':libraries:test'
 include ':libraries:splitInstall'
-include ':libraries:theme'
+include ':libraries:designsystem'
 include ':libraries:navigation'
